Refactored flowc
